% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_individual_signature_accuracy.R
\name{compute_individual_signature_accuracy}
\alias{compute_individual_signature_accuracy}
\title{Compute Individual Signature Accuracy}
\usage{
compute_individual_signature_accuracy(simulated_data)
}
\arguments{
\item{simulated_data}{The data object output by simulate_samples()
with exposures calculated using one of the
signature exposure calculation methods.}
}
\value{
A data frame with each signature's accuracy
}
\description{
Each n-of-1 mutation signature method computes the exposure of each
signature (its contribution to the mutation burden). These numeric
outcomes are stored in the "exposure" column of the
calculated_exposures table in simulated data output. In order to
evaluate each method's accuracy at assigning exposure values, this method
uses various distance metrics to compare the true signatures which gave 
rise to a simulated mutation with the reconstructed signatures estimated 
by the signature method.
}
\details{
Unlike compute_quantitative_accuracy(), this function computes the individual
accuracy of each signature. This allows one to estimate how accurately each
individual signature is predicted with a given method.
}
