% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quantitative_accuracy.R
\name{compute_quantitative_accuracy}
\alias{compute_quantitative_accuracy}
\title{Compute Quantitative Accuracy Score}
\usage{
compute_quantitative_accuracy(simulated_data)
}
\arguments{
\item{simulated_data}{The data object output by simulate_samples()
with exposures calculated using one of the
signature exposure calculation methods.}
}
\value{
A list with named items that report the distance metrics.
}
\description{
Each n-of-1 mutation signature method computes the exposure of each
signature (its contribution to the mutation burden). These numeric
outcomes are stored in the "exposure" column of the
calculated_exposures table in simulated data output. In order to
evaluate each method's accuracy at assigning exposure values, this method
uses various distance metrics to compare the true signatures which gave 
rise to a simulated mutation with the reconstructed signatures estimated 
by the signature method.
}
\details{
The metrics reported are L2 normalized Euclidean distance and cosine distance.
In L2 normalized Euclidean distance, the exposure vectors \eqn{\mathbf{x}, \mathbf{y}} are
normalized as in \eqn{\mathbf{x'} = frac{\mathbf{x}}{\lVert \mathbf{x} \rVert}} into
unit vectors. Then the Euclidean distance \eqn{\lVert \mathbf{y'} - \mathbf{x'} \rVert} is computed.

The cosine distance is calculated as
\eqn{1 - \frac{\mathbf{x} \cdot \mathbf{y}}{\lVert \mathbf{x} \rVert \lVert \mathbf{y} \rVert}}.
}
