% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstructSigs_runner.R
\name{deconstructSigs_runner}
\alias{deconstructSigs_runner}
\title{Computes exposures by the deconstructSigs method}
\usage{
deconstructSigs_runner(simulated_data)
}
\arguments{
\item{simulated_data}{Simulated data object obtained by
\code{\link{simulate_signatures}}.}
}
\value{
A list containing a table of signature exposures and the name of the method used.
}
\description{
deconstructSigs is a method for computing n-of-1 mutation signatures
reported in Genome Biology 
(https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-0893-4)
and implemented as a package on CRAN 
(https://cran.rstudio.com/web/packages/deconstructSigs/index.html).
}
\details{
This method runs deconstructSigs on a simulated data object and
returns the associated signature exposures. Because deconstructSigs
employs its own feature selection, any signature with a non-zero
exposure value is considered to be a contributor to the mutation burden.
Thus, this method calls a signature present if there is at least
one mutation contributed by that signature.
}
