% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_samples.R
\name{simulate_samples}
\alias{simulate_samples}
\title{Simulate samples from reference signatures}
\usage{
simulate_samples(reference_signatures = wtsi_30_snv_signatures,
  n_mutations = 1000, n_active_signatures = 5, chosen_signatures = c(),
  exposure_sampler = uniform_exposure_sampler,
  perturbation_percent_deviation = 0)
}
\description{
Simulate samples from reference signatures
}
